move Zmmumerge to use DiMuonMassProfiles
